Fix evaluation by exporting in the original size
previously, when the prostate cropping was added, the images during
export to nifti (see segmentation_export.py) were exported cropped
(but with correct affine).

This was problematic during the evaluation, which is based on the
niftis and the original shape.

Thus, the cropped area is re-introduced, by modifying the crop_bbox,
so that the exported niftis now have the original shape.